refactor: gpt 통신용 dto 와 구분을 위한 이름 변경
