+++ src/main/java/com/kuit/agarang/domain/ai/controller/MemoryAIController.java
@@ -1,6 +1,6 @@
 package com.kuit.agarang.domain.ai.controller;
 
-import com.kuit.agarang.domain.ai.model.dto.gpt.GPTQuestionResponse;
+import com.kuit.agarang.domain.ai.model.dto.QuestionResponse;
 import com.kuit.agarang.domain.ai.service.MemoryAIService;
 import com.kuit.agarang.global.common.response.BaseResponse;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +19,7 @@ public class MemoryAIController {
   private final MemoryAIService memoryAIService;
 
   @PostMapping("/image-to-question")
-  public ResponseEntity<BaseResponse<GPTQuestionResponse>> getImageQuestion(@RequestParam MultipartFile image) throws Exception {
+  public ResponseEntity<BaseResponse<QuestionResponse>> getImageQuestion(@RequestParam MultipartFile image) throws Exception {
     return ResponseEntity.ok(new BaseResponse<>(memoryAIService.getFirstQuestion(image)));
   }
 }
