+++ src/main/java/com/kuit/agarang/domain/ai/model/dto/QuestionResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.ai.model.dto.gpt;
+package com.kuit.agarang.domain.ai.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class GPTQuestionResponse {
-  private GPTQuestionResult question;
+public class QuestionResponse {
+  private QuestionResult question;
 }
