+++ src/main/java/com/kuit/agarang/domain/ai/model/dto/QuestionResult.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.ai.model.dto.gpt;
+package com.kuit.agarang.domain.ai.model.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,13 +6,13 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class GPTQuestionResult {
+public class QuestionResult {
   private String id;
   private String text;
   private String audioUrl;
 
   @Builder
-  public GPTQuestionResult(String id, String text, String audioUrl) {
+  public QuestionResult(String id, String text, String audioUrl) {
     this.id = id;
     this.text = text;
     this.audioUrl = audioUrl;
